Reestruturando diretorios de acordo com a extensao das regras, percorrendo todas as regras
