+++ bin/moenda.js
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 
 const util = require('../src/util');
-const os = require('os');
 const {Command} = require('commander');
 const program = new Command();
+const genericRules = require("../src/rules/genericRules");
 program.version('0.0.1');
 
 program
@@ -15,8 +15,8 @@ program
 program.parse(process.argv);
 
 function main() {
-  if (util.testIfIsFile(program.path).status === true) {
-    util.exitAid(program);
+  if (util.testIfIsFile(program.path) === true) {
+    util.exitAid(program, genericRules);
   } else {
     let parameter = '';
 
@@ -27,7 +27,7 @@ function main() {
     const array = util.directoryFiles(program.path, parameter).data;
 
     for (let i = 0; i < array.length; i++) {
-      util.exitAid(program, array[i]);
+      util.exitAid(program, genericRules, array[i]);
     }
   }
 }
