+++ src/rules/genericRules.js
@@ -1,9 +1,13 @@
+
 const fs = require('fs');
 const os = require('os');
-const util = require('./util');
+
+
 
 module.exports = {
+  
   ruleLineCounter: function (link, rules) {
+    const util = require('../util');
     let result = {
       line: '-',
       column: '-',
@@ -34,6 +38,7 @@ module.exports = {
   },
 
   ruleLessThanXLines: function (link, rules) {
+    const util = require('../util');
     let result = {
       line: '-',
       column: '-',
@@ -43,7 +48,7 @@ module.exports = {
       toString: '',
       name: 'ruleLessThanXLines()',
     };
-
+    
     if (util.testIfIsFile(link) === true) {
       const lines = fs.readFileSync(link, 'utf-8').split(os.EOL).length;
       result.msg = `The file has ${lines} line(s)`;
@@ -51,6 +56,7 @@ module.exports = {
 
       if (rules.ruleLessThanXLines?.limit !== undefined) {
         if (lines < rules.ruleLessThanXLines.limit) {
+          
           result.status = 'error';
           result.line = lines;
           result.msg = `This file is expected to have a minimum of ${rules.ruleLessThanXLines.limit} lines`;
@@ -64,6 +70,7 @@ module.exports = {
   },
 
   ruleLineAboveXCharacters: function (link, rules) {
+    const util = require('../util');
     const fileTest = fs.statSync(link, 'utf-8');
     let result = {
       status: false,
@@ -79,12 +86,13 @@ module.exports = {
       if (rules.ruleLineAboveXCharacters?.limit !== undefined) {
         const file = fs.readFileSync(link, 'utf-8');
         const lines = file.split(os.EOL);
+
         for (let i = 0; i < lines.length; i++) {
           if (lines[i].length > rules.ruleLineAboveXCharacters.limit) {
             result.status = 'error';
             result.line = i + 1;
             result.data = lines[i];
-            result.column = rules.LineAboveXCharacters.limit + 1;
+            result.column = rules.ruleLineAboveXCharacters.limit + 1;
             result.msg = `This line must not exceed ${rules.ruleLineAboveXCharacters.limit} characters.`;
             result.name = 'ruleLineAboveXCharacters()';
             result.toString = util.toStringGenerate(result);
@@ -97,112 +105,8 @@ module.exports = {
     return result;
   },
 
-  ruleFirstSectionStartsWithHx: function (link, rules) {
-    let result = {
-      status: '',
-      line: '-',
-      column: '-',
-      msg: '',
-      data: '',
-      toString: '',
-      name: '',
-    };
-
-    if (util.testIfIsFile(link) === true && link.endsWith('.md')) {
-      const file = fs.readFileSync(link, 'utf-8');
-      const lines = file.split(os.EOL);
-      let size = [];
-
-      if (rules.ruleFirstSectionStartsWithHx?.limit !== undefined) {
-        result = {
-          status: 'error',
-          line: '-',
-          column: '-',
-          msg: `The first section of the file does not start with H${rules.ruleFirstSectionStartsWithHx.limit}`,
-          data: '',
-          toString: '',
-          name: '',
-        };
-        for (let i = 0; i < lines.length; i++) {
-          if (lines[i].startsWith('#')) {
-            let aux = '';
-
-            for (let j = 0; j < lines[i].length; j++) {
-              if (lines[i][j] === '#') {
-                aux += '#';
-              }
-            }
-            size[size.length] = [i, aux.length];
-            result.column = aux.length;
-            result.data = aux;
-            result.name = 'ruleFirstSectionStartsWithH1()';
-
-            result.line = size[0][0] + 1;
-
-            if (size[0][1] === rules.ruleFirstSectionStartsWithHx.limit) {
-              result.status = false;
-            }
-            break;
-          }
-        }
-      }
-    }
-    result.toString = util.toStringGenerate(result);
-
-    return result;
-  },
-
-  ruleNeighboringSections: function (link, config) {
-    let result = {
-      status: false,
-      line: '-',
-      column: '-',
-      msg: '',
-      data: '',
-      toString: '',
-      name: '',
-    };
-
-    if (util.testIfIsFile(link) === true && link.endsWith('.md')) {
-      const file = fs.readFileSync(link, 'utf-8');
-      const lines = file.split(os.EOL);
-      let sizes = [];
-
-      for (let i = 0; i < lines.length; i++) {
-        if (lines[i].startsWith('#')) {
-          let aux = '';
-          for (let j = 0; j < lines[i].length; j++) {
-            if (lines[i][j] === '#') {
-              aux += '#';
-            }
-          }
-          sizes[sizes.length] = [i, aux.length];
-        }
-      }
-
-      for (let i = 0; i < sizes.length - 1; i++) {
-        if (
-          sizes[i][1] !== sizes[i + 1][1] &&
-          sizes[i][1] !== sizes[i + 1][1] - 1 &&
-          sizes[i][1] !== sizes[i + 1][1] + 1
-        ) {
-          result.status = 'error';
-          result.line = sizes[i][0] + 1;
-          result.msg =
-            'The file does not have increasing and decreasing sections';
-          result.column = sizes[i][1];
-          result.data = lines[sizes[i][0]];
-          result.name = 'ruleNeighboringSections()';
-        }
-      }
-    }
-
-    result.toString = util.toStringGenerate(result);
-
-    return result;
-  },
-
   ruleInconsistencyOfSpaces: function (path, config) {
+    const util = require('../util');
     let result = {
       status: false,
       line: '-',
@@ -254,6 +158,7 @@ module.exports = {
   },
 
   ruleConsecutiveBlankLines: function (link, rules) {
+    const util = require('../util');
     let result = {
       line: '-',
       column: '-',
